Revert "Move search"
This reverts commit 6c2f0c5e2f1db227f61457b01970d4a79566e43f.
