+++ crates/swc_common/src/util/mod.rs
@@ -1,5 +1,4 @@
 pub mod iter;
 pub mod map;
 pub mod move_map;
-pub mod search;
 pub mod take;
